back

+++ frontend/script.js
@@ -81,28 +81,6 @@ addAluno.addEventListener("click", function(e) {
     }
 });
 
-
-/* ao apertar verAlunos, para cada aluno recebido de localhost:3058/alunos, dentro de alunosList, insira: */
-/*
-<table>
-    <hr>
-    <tr>
-        <th class="headerTH">Nome</th>
-        <td id="nomeAluno" class="infoAluno">${aluno.nome} ${aluno.apelido}</th>
-    </tr>
-    <tr>
-        <th class="headerTH">Curso</th>
-        <td id="cursoAluno" class="infoAluno">${fixCurso(aluno.curso)}</th>
-    </tr>
-    <tr>
-        <th class="headerTH">Ano</th>
-        <td id="anoAluno" class="infoAluno">${fixAno(aluno.ano)}</th>
-        </tr>
-</table>
-<button id="delAluno">Apagar ${aluno.nome}</button>
-            
-*/
-
 verAlunos.addEventListener("click", function() {
     alunosList.innerHTML = "";
     fetch("http://localhost:3058/alunos")
@@ -121,11 +99,11 @@ verAlunos.addEventListener("click", function() {
                         </tr>
                         <tr>
                             <th class="headerTH">Curso</th>
-                            <td id="cursoAluno" class="infoAluno">${(aluno.curso)}</td>
+                            <td id="cursoAluno" class="infoAluno">${fixCurso(aluno.curso)}</td>
                         </tr>
                         <tr>
                             <th class="headerTH">Ano</th>
-                            <td id="anoAluno" class="infoAluno">${(aluno.anoCurricular)}</td>
+                            <td id="anoAluno" class="infoAluno">${fixAno(aluno.anoCurricular)}</td>
                         </tr>
                     </table>
                     <button class="delAluno btnRed" id="${aluno._id}">Apagar ${aluno.nome}</button>
@@ -134,17 +112,20 @@ verAlunos.addEventListener("click", function() {
                 alunosList.appendChild(alunoDiv);
             });
 
-            // Adiciona o evento de exclusão para cada botão "Apagar"
-            document.querySelectorAll("#delAluno").forEach(button => {
+            const delAlunos = document.querySelectorAll(".delAluno");
+            delAlunos.forEach(button => {
                 button.addEventListener("click", function() {
-                    const id = this.getAttribute("data-id");
+                    if (!confirm("Tem certeza que deseja apagar este aluno? Esta ação não pode ser desfeita.")) {
+                        return;
+                    }
+                    const id = this.id;
                     fetch(`http://localhost:3058/alunos/${id}`, {
                         method: "DELETE"
                     })
                     .then(response => response.json())
                     .then(data => {
-                        console.log("Aluno deletado:", data);
-                        verAlunos.click(); // Atualiza a lista de alunos
+                        console.log(data);
+                        verAlunos.click();
                     })
                     .catch(error => {
                         console.error("Erro ao deletar aluno:", error);
@@ -156,3 +137,33 @@ verAlunos.addEventListener("click", function() {
             console.error("Erro ao buscar alunos:", error);
         });
 });
+
+function fixCurso(curso) {
+    switch (curso) {
+        case "1":
+            return "Engenharia da Computação Gráfica e Multimédia";
+        case "2":
+            return "Engenharia de Redes e Sistemas de Computadores";
+        case "3":
+            return "Bacharelado em Sistemas de Informação";
+        case "4":
+            return "Curso de Graduação em Cinema";
+        default:
+            return curso;
+    }
+}
+
+function fixAno(ano) {
+    switch (ano) {
+        case "1":
+            return "1º Ano";
+        case "2":
+            return "2º Ano";
+        case "3":
+            return "3º Ano";
+        case "4":
+            return "4º Ano";
+        default:
+            return ano;
+    }
+}
